nvim: lua the augroup NvimTreeAutoClose (!!)
There is an inconsistency of behaviour in Vimcript. If I open a file and
later open nvim-tree, closing the file window does not auto close the
NvimTree buffer. I need to close the buffer using `q` key, and then do
the above steps again to get the effect.

In Lua, however, the NvimTree buffer will not be closed no matter what.

This could be the problem of Neovim 0.7.0 itself